some changes
change something

+++ DataGenerationTool/src/com/ligadata/dataGenerationTool/TimeUtility.java
@@ -1,8 +1,5 @@
 package com.ligadata.dataGenerationTool;
 
-import java.text.SimpleDateFormat;
-import java.util.Date;
-
 import com.ligadata.dataGenerationToolBean.ConfigObj;
 import com.ligadata.dataGenerationToolBean.FileNameConfig;
 
@@ -56,12 +53,12 @@ public class TimeUtility {
 		
 		endTime = currentTime + ( multiplyFactor * timeAmountForFileSplit);	
 				
-		if ((long)fileNameConfig.getNextFileTime() == 0) {
+		if (fileNameConfig.getNextFileTime() == 0) {
 			fileNameConfig.setNextFileTime(endTime);
 		}
 
 
-		if ( (long)currentTime >= (long)fileNameConfig.getNextFileTime()) {   //  Double.compare(currentTime,FilesUtility.nextFileTime ) < 0
+		if ( currentTime >= fileNameConfig.getNextFileTime()) {   //  Double.compare(currentTime,FilesUtility.nextFileTime ) < 0
 			fileNameConfig.setOldFileTime(fileNameConfig.getNextFileTime());
 			fileNameConfig.setNextFileTime(endTime);
 			return true;
